+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/NoEnumTrailingCommaCheckTest.java
@@ -34,7 +34,7 @@ public class NoEnumTrailingCommaCheckTest extends AbstractModuleTestSupport {
     }
 
     @Test
-    public void testDefault() throws Exception {
+    public void testDefaultOne() throws Exception {
         final String[] expected = {
             "23:12: " + getCheckMessage(MSG_KEY),
             "28:12: " + getCheckMessage(MSG_KEY),
@@ -44,23 +44,39 @@ public class NoEnumTrailingCommaCheckTest extends AbstractModuleTestSupport {
             "53:21: " + getCheckMessage(MSG_KEY),
             "58:12: " + getCheckMessage(MSG_KEY),
             "76:9: " + getCheckMessage(MSG_KEY),
-            "106:55: " + getCheckMessage(MSG_KEY),
-            "110:14: " + getCheckMessage(MSG_KEY),
-            "114:14: " + getCheckMessage(MSG_KEY),
-            "130:14: " + getCheckMessage(MSG_KEY),
-            "134:14: " + getCheckMessage(MSG_KEY),
-            "138:55: " + getCheckMessage(MSG_KEY),
-            "144:33: " + getCheckMessage(MSG_KEY),
-            "151:33: " + getCheckMessage(MSG_KEY),
-            "168:13: " + getCheckMessage(MSG_KEY),
-            "186:21: " + getCheckMessage(MSG_KEY),
-            "206:10: " + getCheckMessage(MSG_KEY),
-            "250:55: " + getCheckMessage(MSG_KEY),
-            "255:9: " + getCheckMessage(MSG_KEY),
-            "258:9: " + getCheckMessage(MSG_KEY),
         };
         verifyWithInlineConfigParser(
-                getPath("InputNoEnumTrailingComma.java"), expected);
+                getPath("InputNoEnumTrailingCommaOne.java"), expected);
+    }
+
+    @Test
+    public void testDefaultTwo() throws Exception {
+        final String[] expected = {
+            "20:55: " + getCheckMessage(MSG_KEY),
+            "24:14: " + getCheckMessage(MSG_KEY),
+            "28:14: " + getCheckMessage(MSG_KEY),
+            "44:14: " + getCheckMessage(MSG_KEY),
+            "48:14: " + getCheckMessage(MSG_KEY),
+            "52:55: " + getCheckMessage(MSG_KEY),
+            "58:33: " + getCheckMessage(MSG_KEY),
+            "65:33: " + getCheckMessage(MSG_KEY),
+            "82:13: " + getCheckMessage(MSG_KEY),
+        };
+        verifyWithInlineConfigParser(
+                getPath("InputNoEnumTrailingCommaTwo.java"), expected);
+    }
+
+    @Test
+    public void testDefaultThree() throws Exception {
+        final String[] expected = {
+            "13:21: " + getCheckMessage(MSG_KEY),
+            "33:10: " + getCheckMessage(MSG_KEY),
+            "77:55: " + getCheckMessage(MSG_KEY),
+            "82:9: " + getCheckMessage(MSG_KEY),
+            "85:9: " + getCheckMessage(MSG_KEY),
+        };
+        verifyWithInlineConfigParser(
+                getPath("InputNoEnumTrailingCommaThree.java"), expected);
     }
 
     @Test
